requirements.txt
aiohttp==3.9.5
pillow==10.3.0
python-dotenv==1.0.1
tqdm==4.66.4
aiofiles==23.2.1
flask==3.0.3
